tests/integration/apps/kafka-app/requirements.txt
starlette
uvicorn
kafka-python==2.0.2
confluent-kafka
